# SteticOpenApi.yaml
openapi: 3.0.1
info:
  title: Stetic API
  description: API para la gestión de un centro de estética
  version: 1.0.0
servers:
  - url: 'https://tiendasphy.000webhostapp.com/index.php'
...
paths:
...
  /products:
    get:
      tags:
        - products
      summary: Lista todos los productos o filtra por búsqueda
      parameters:
        - name: name
          in: query
          description: Nombre del producto
          required: false
          schema:
            type: string
        - name: productId
          in: query
          description: ID del producto
          required: false
          schema:
            type: integer
        - name: dangerous
          in: query
          description: producto peligroso
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Devuelve la lista completa de productos o los filtrados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    post:
      tags:
        - products
      summary: Añade un nuevo producto
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: El producto se ha registrado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Error de validación en los datos de entrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /product/{productId}:
    get:
      tags:
        - products
      summary: Obtiene un producto por ID
      parameters:
        - name: productId
          in: path
          description: ID del producto
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Devuelve el producto que coincide con el ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'

    delete:
      tags:
        - products
      summary: Elimina un producto por ID
      parameters:
        - name: productId
          in: path
          description: ID del producto
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: El producto se ha eliminado correctamente
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'

    put:
      tags:
        - products
      summary: Modifica un producto por ID
      parameters:
        - name: productId
          in: path
          description: ID del producto
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: El producto se ha modificado correctamente
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        registrationDate:
          type: string
          format: date
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        errors:
          type: array
          items:
            type: string
    InternalServerError:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
    BadRequest:
      type: object
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
      example:
        code: 400
        message: Validation error
        errors:
          - El nombre es obligatorio
          - El precio debe ser mayor que cero